requirements.txt
# local package
-e .

# external requirements
torch>=1.8.1
torchvision>=0.9.1
matplotlib>=3.3.4
kornia>=0.5.3
gdown>=3.13.0
azureml-core>=1.30.0
flake8
pillow
pandas
numpy
click
scikit-learn
